linux

+++ scripts/fetch_libvlc.ps1
@@ -3,6 +3,11 @@ param(
 )
 
 $win64Url = "https://artifacts.videolan.org/vlc/nightly-win64/20260615-0427/vlc-4.0.0-dev-win64-853c0f19.7z"
+$linux64Url = "https://github.com/xiSage/libvlc-4-linux-binary/releases/download/vlc_4.0.0-dev-37536-g546e18e53e_amd64/linux-x64.tar.gz"
+
+if (!(Test-Path "thirdparty\vlc")) {
+    New-Item -Path "thirdparty\vlc" -ItemType Directory -Force | Out-Null
+}
 
 if ($Platforms -contains "win64") {
     $archive = "vlc-win64.7z"
@@ -31,5 +36,17 @@ if ($Platforms -contains "win64") {
 
     # Cleanup
     Remove-Item -Recurse -Force $tempDir
+    Remove-Item -Force $archive
+}
+
+if ($Platforms -contains "linux64") {
+    $archive = "vlc-linux64.tar.gz"
+    $targetDir = "thirdparty/vlc"
+
+    Invoke-WebRequest -Uri $linux64Url -OutFile $archive
+
+    tar -zxvf"$archive" -C $targetDir
+
+    # Cleanup
     Remove-Item -Force $archive
 }
